agent/pyproject.toml
[project]
name = "ha-agent"
version = "0.1.0"
description = "Home Assistant ADK-based action-taking agent"
requires-python = ">=3.12,<3.13"
readme = "README.md"
dependencies = [
    "google-adk~=1.31.0",
    "fastapi~=0.136.0",
    "uvicorn[standard]~=0.44.0",
    "pydantic~=2.13.0",
    "pydantic-settings~=2.13.0",
    "httpx~=0.28.0",
    "websockets~=15.0",
    "sentence-transformers~=5.4.0",
    "numpy~=2.4.0",
    "sse-starlette~=3.3.0",
    "python-json-logger~=4.1.0",
]

[dependency-groups]
dev = [
    "pytest~=9.0.0",
    "pytest-asyncio~=1.3.0",
    "pytest-httpx~=0.36.0",
    "freezegun~=1.5.0",
    "ruff~=0.15.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ha_agent"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "RUF"]

[tool.ruff.lint.isort]
known-first-party = ["ha_agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
pythonpath = ["src"]
